DEML framework with multiple meta machine learning models

Stacked ensemble models with multiple meta machine learning models

stack_ensemble(
  object,
  meta_model = c("SL.ranger"),
  original_feature = FALSE,
  X,
  ...
)

Arguments

object

the object that predictModel function created

meta_model

character string in format of "SL.xgboost","SL.randomForest" as the stack ensemble models

original_feature

a Boolean value, whether integrate the original feature into the meta model to build the stack ensemble models.

X

optional when 'original_feature = FALSE', the predictors (eg. temp, wind speed) in training data that we

...

other paramaters that belong to 'SuperLearner' pacakge

Value

a list with components: the information about base model; the information about meta model; ensemble_pre (the ensemble model predict value) when object has the newX parameter; stack_ensemble_pred (the double stack ensemble model predict value) when object has the newX parameter; the DEML model assessment(R-squire and RMSE)

Examples

if (FALSE) {
# Let's predict the PM2.5 values with machine learming methods

# Splitting training set into two parts based on outcome: 80% and 20%
set.seed(1234)
data("envir_example")
size <-
  caret::createDataPartition(y = envir_example$PM2.5, p = 0.8, list = FALSE)
trainset <- envir_example[size, ]
testset <- envir_example[-size, ]

# set the predictor "PM2.5" and dependence variables
y <- c("PM2.5")
x <- colnames(envir_example[-c(1, 6)]) # except "date" and "PM2.5"

# object do not include newX dataset
pre_stack_2 <-
  stack_ensemble(
    object = pred_m2,
    meta_model = c("SL.ranger", "SL.xgboost", "SL.lm"),
    original_feature = FALSE,
    X = testset[, x]
  )

# prediction
pred_stack_2_new <- predict(object = pre_stack_2, newX = testset[, x])
}